refactor(dashboard): move all filesystem settings into one new page
